requirements.txt
yml
pydantic
python-dotenv
openai
cohere
anthropic
huggingface_hub
protobuf
tenacity
